changed long to int32_t to make sure strict testcase8 pass and mem limit pass

+++ _includes/source_code/code/24-PDP/souvlakia/souvlakia_nlogn_implicit_st.cc
@@ -8,6 +8,7 @@
 #include <cstdio>
 #include <algorithm>
 #include <climits>
+#include <cstdint>
 #include <cassert>
 #include <vector>
 #include <list>
@@ -18,18 +19,18 @@
 
 using namespace std;
 
-const long 
-	MAXN = long(1e5),	//max shops
-	MAXQ = long(5e4),	//max queries
-	INF  = LONG_MAX;	//INF for dijkstra
-long	N,M,Q,C[3];
+const int32_t 
+	MAXN = int32_t(1e5),	//max shops
+	MAXQ = int32_t(5e4),	//max queries
+	INF  = INT32_MAX;	//INF for dijkstra
+int32_t	N,M,Q,C[3];
 
 struct shop {
 #define	X	d[0]
 #define Y 	d[1]
 #define Z	d[2]
-	long d[3];//The 3 distances (X,Y,Z) from {A,B,C}
-	list<long> queryid;//query ids that has to answer shop[i]
+	int32_t d[3];//The 3 distances (X,Y,Z) from {A,B,C}
+	list<int32_t> queryid;//query ids that has to answer shop[i]
 			//use list in case there are more than one queries per shop
 	shop(){ X=Y=Z=INF; }//help dijsktra initial values
 	bool operator <(const shop& b) const { 
@@ -37,10 +38,10 @@ struct shop {
 	}
 } S[MAXN+1];
 
-void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
+void dijsktra(int32_t src,int dupd,vector<vector<pair<int32_t,int32_t>>>& edge){
 	//src is one of {A,B,C}
 	//dupd in [0,3). Update Shop[*].d[dupd] distance from src
-	set<pair<long,long>> DS;//<distance,shopid> like a priority queue
+	set<pair<int32_t,int32_t>> DS;//<distance,shopid> like a priority queue
 	
 	DS.insert({0,src});
 	S[src].d[dupd] = 0;
@@ -48,9 +49,9 @@ void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
 	while(!DS.empty()){//explore
 		auto x = *(DS.begin());//head of priority queue
 		DS.erase(DS.begin());
-		long daddy = x.second;
+		int32_t daddy = x.second;
 		for(auto y : edge[daddy]){//expand graph
-			long child = y.first, dist = y.second;
+			int32_t child = y.first, dist = y.second;
 			if(S[daddy].d[dupd]!=INF && S[child].d[dupd] > S[daddy].d[dupd] + dist){
 				//this is an improvement
 				if(S[child].d[dupd]!=INF)//was it in the queue already?
@@ -62,8 +63,8 @@ void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
 	}
 }
 
-inline long read_fast(){//fast IO is required for this problem due to huge io
-	long x = 0;
+inline int32_t read_fast(){//fast IO is required for this problem due to huge io
+	int32_t x = 0;
 	char c;
 	while((c=getchar_unlocked())<'0');//skip blank,CR,TF,TAB
 	do
@@ -81,16 +82,16 @@ inline void write_fast(bool f){
 //RMQ with implicit segment tree node and functions
 //segment tree nodes are created upon request (by expand function)
 //segment tree usage is RangeMinimumQuery for Y in [0,20000*MAXN] range
-long stn;//keep track of ST[] pool usage
-const long SMAX = MAXN * 30;//log2(MAXN * D) == 30
-const long YMAX = long(2e4)*MAXN;	//max value of Y
+int32_t stn;//keep track of ST[] pool usage
+const int32_t SMAX = MAXN * 30;//log2(MAXN * D) == 30
+const int32_t YMAX = int32_t(2e4)*MAXN;	//max value of Y
 
 struct node {//implicit segment tree (lazy build)
-	long left,right;//Y limits that this node handles
-	long minZ;//min Shop.Z value of segment tree
-	long leftptr,rightptr;//"pointers" to the children 
+	int32_t left,right;//Y limits that this node handles
+	int32_t minZ;//min Shop.Z value of segment tree
+	int32_t leftptr,rightptr;//"pointers" to the children 
 		
-	node(long le,long ri,long z){
+	node(int32_t le,int32_t ri,int32_t z){
 		left = le, right = ri, minZ = z;
 		leftptr = rightptr = 0;
 	}
@@ -100,16 +101,16 @@ struct node {//implicit segment tree (lazy build)
 void sinit(){
 	st[0] = node(0,YMAX,INF);//root node
 }
-void expand(long root){
+void expand(int32_t root){
 	node& s = st[root];
 	if(!s.leftptr && s.left<s.right){
-		long mid = (s.left + s.right)/2;
+		int32_t mid = (s.left + s.right)/2;
 		assert(stn + 2 < SMAX);
 		st[s.leftptr = ++stn] = node(s.left,mid,s.minZ);
 		st[s.rightptr= ++stn] = node(mid+1,s.right,s.minZ);
 	}
 }
-long squery(long y,long root = 0){
+int32_t squery(int32_t y,int32_t root = 0){
 	node& s = st[root];
 	if(s.right <= y)//completely involved
 		return s.minZ;
@@ -119,7 +120,7 @@ long squery(long y,long root = 0){
 	assert(s.leftptr);assert(s.rightptr);
 	return min(squery(y,s.leftptr),squery(y,s.rightptr));
 }
-void supdate(long y,long z,long root = 0){
+void supdate(int32_t y,int32_t z,int32_t root = 0){
 	node& s = st[root];	
 	s.minZ = min(s.minZ,z);
 	if(s.right == y)
@@ -141,13 +142,13 @@ int main(){
 	N = read_fast(), M = read_fast();
 
 	{//hint: edge vector below is temporary
-		vector<vector<pair<long,long>>> edge(N+1);//<edge_to, distance>
-		for(long a,b,d,i=0;i<M;++i){
+		vector<vector<pair<int32_t,int32_t>>> edge(N+1);//<edge_to, distance>
+		for(int32_t a,b,d,i=0;i<M;++i){
 			a = read_fast(), b = read_fast(), d = read_fast();
 			edge[a].push_back({b,d});
 			edge[b].push_back({a,d});
 		}
-		for(long i=0;i<3;i++){
+		for(int i=0;i<3;i++){
 			C[i]=read_fast();
 			dijsktra(C[i],i,edge);
 		}
@@ -155,7 +156,7 @@ int main(){
 	}//memory of edge vector is released
 
 	//read queries and store reference so we can answer offline
-	for(long i=0;i<Q;++i)
+	for(int32_t i=0;i<Q;++i)
 		S[read_fast()].queryid.push_back(i);//shop[q] has to answer i-th query
 
 	sort(S+1,S+N+1);
@@ -163,7 +164,7 @@ int main(){
 	vector<bool> ans(MAXQ+1,false);		//offline answers
 
 	sinit();
-	for(long right = 1, left = 1; left <=N;){
+	for(int32_t right = 1, left = 1; left <=N;){
 		while(right<=N && S[right].X == S[left].X){
 			//for all shops with equal X postpone updates
 			const shop& s = S[right];
@@ -179,7 +180,7 @@ int main(){
 		}
 	}
 	
-	for(long i=0;i<Q;++i)
+	for(int32_t i=0;i<Q;++i)
 		write_fast(ans[i]);
 	
 	return 0;
