+++ _includes/source_code/code/24-PDP/souvlakia/souvlakia_nlogn_map.cc
@@ -5,6 +5,7 @@
 #include <cstdio>
 #include <algorithm>
 #include <climits>
+#include <cstdint>
 #include <cassert>
 #include <vector>
 #include <list>
@@ -15,10 +16,10 @@
 
 using namespace std;
 
-const long 
-	MAXN = long(1e5),	//max shops
-	MAXQ = long(5e4),	//max queries
-	INF  = LONG_MAX;	//INF for dijkstra
+const int32_t
+	MAXN = int32_t(1e5),	//max shops
+	MAXQ = int32_t(5e4),	//max queries
+	INF  = INT32_MAX;	//INF for dijkstra
 	
 long	N,M,Q,C[3];
 
@@ -26,8 +27,8 @@ struct shop {
 #define	X	d[0]
 #define Y 	d[1]
 #define Z	d[2]
-	long d[3];//The 3 distances (X,Y,Z) from {A,B,C}
-	list<long> queryid;//query ids that has to answer shop[i]
+	int32_t d[3];//The 3 distances (X,Y,Z) from {A,B,C}
+	list<int32_t> queryid;//query ids that has to answer shop[i]
 			//use list in case there are more than one queries per shop
 	shop(){ X=Y=Z=INF; }//help dijsktra initial values
 	bool operator <(const shop& b) const { 
@@ -35,10 +36,10 @@ struct shop {
 	}
 } S[MAXN+1];
 
-void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
+void dijsktra(int32_t src,int dupd,vector<vector<pair<int32_t,int32_t>>>& edge){
 	//src is one of {A,B,C}
 	//dupd in [0,3). Update Shop[*].d[dupd] distance from src
-	set<pair<long,long>> DS;//<distance,shopid> like a priority queue
+	set<pair<int32_t,int32_t>> DS;//<distance,shopid> like a priority queue
 	
 	DS.insert({0,src});
 	S[src].d[dupd] = 0;
@@ -46,9 +47,9 @@ void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
 	while(!DS.empty()){//explore
 		auto x = *(DS.begin());//head of priority queue
 		DS.erase(DS.begin());
-		long daddy = x.second;
+		int32_t daddy = x.second;
 		for(auto y : edge[daddy]){//expand graph
-			long child = y.first, dist = y.second;
+			int32_t child = y.first, dist = y.second;
 			if(S[daddy].d[dupd]!=INF && S[child].d[dupd] > S[daddy].d[dupd] + dist){
 				//this is an improvement
 				if(S[child].d[dupd]!=INF)//was it in the queue already?
@@ -60,8 +61,8 @@ void dijsktra(long src,long dupd,vector<vector<pair<long,long>>>& edge){
 	}
 }
 
-inline int read_fast(){//fast IO is required for this problem due to huge io
-	int x = 0;
+inline int32_t read_fast(){//fast IO is required for this problem due to huge io
+	int32_t x = 0;
 	char c;
 	while((c=getchar_unlocked())<'0');//skip blanks, CR, LF
 	do
@@ -80,8 +81,8 @@ inline void write_fast(bool f){
 //sequence of Y line, the corresponding descending sequence of Z values. 
 //In fact forevery Y<Y' <=> Z>Z' (stricly descreasing subsequence)
 //So let's make it in a map of Y values to Z
-map<long,long> Zmin;
-typedef map<long,long>::iterator iter;
+map<int32_t,int32_t> Zmin;
+typedef map<int32_t,int32_t>::iterator iter;
 
 void sinit(){
 	//add some edge limits so we'll not have to check for 
@@ -89,12 +90,12 @@ void sinit(){
 	Zmin.insert({-2,INF});//lower limit for optimization and queries
 	Zmin.insert({INF,-1});//upper limit for optimizations
 }
-long squery(long y){
+long squery(int32_t y){
 	iter it = Zmin.lower_bound(y-1);
 	while(it->first >= y)it--;
 	return it->second;
 }
-void supdate(long y,long z){
+void supdate(int32_t y,int32_t z){
 	iter curr = Zmin.find(y);
 	if(curr == Zmin.end()){
 		curr = (Zmin.insert({y,z})).first;
@@ -131,8 +132,8 @@ int main(){
 		//max mem usage: 400.000 * 2 * 3 * sizeof(long) = 2.400.000 * sizeof(long) = 19.2Mb
 		//and we definitely benefit our program by releasing this mem when no more required
 		//if we use a dynamic allocated scheme for implicit segment tree or use map template
-		vector<vector<pair<long,long>>> edge(N+1);//<edge_to, distance>
-		for(long a,b,d,i=0;i<M;++i){
+		vector<vector<pair<int32_t,int32_t>>> edge(N+1);//<edge_to, distance>
+		for(int32_t a,b,d,i=0;i<M;++i){
 			a = read_fast(), b = read_fast(), d = read_fast();
 			edge[a].push_back({b,d});
 			edge[b].push_back({a,d});
@@ -141,12 +142,12 @@ int main(){
 			C[i] = read_fast();
 		Q = read_fast();
 
-		for(long i=0;i<3;++i)
+		for(int i=0;i<3;++i)
 			dijsktra(C[i],i,edge);
 	}//memory of edge vector is released
 
 	//read queries and store reference so we can answer offline
-	for(long i=0;i<Q;++i)
+	for(int32_t i=0;i<Q;++i)
 		S[read_fast()].queryid.push_back(i);//shop[q] has to answer i-th query
 
 	sort(S+1,S+N+1);
@@ -154,7 +155,7 @@ int main(){
 	vector<bool> ans(MAXQ+1,false);		//offline answers
 
 	sinit();
-	for(long right = 1, left = 1; left <=N;){
+	for(int32_t right = 1, left = 1; left <=N;){
 		while(right<=N && S[right].X == S[left].X){
 			//for all shops with equal X postpone updates
 			const shop& s = S[right];
@@ -170,7 +171,7 @@ int main(){
 		}
 	}
 	
-	for(long i=0;i<Q;++i)
+	for(int32_t i=0;i<Q;++i)
 		write_fast(ans[i]);
 	return 0;
 }
